examples/pizza-order/pyproject.toml
[project]
name = "pizza-order"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "faker>=37.12.0",
    "fastapi>=0.128.3",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "strawberry-graphql[fastapi]>=0.283.3",
    "uvicorn[standard]>=0.39.0",
]
